requirements.txt
scholarly
requests
beautifulsoup4
pandas
rapidfuzz
groq
python-dotenv
google-search-results